requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
numpy==1.24.3
pandas==2.0.3
typing-extensions>=4.7.0 